lmp/bb-build: only use the sstate-cache mirror on the first step
On the fisrt build step we run bitbake for setscene only tasks
and obvious this one is the only one that use the reads and use the
sstate-cache.
The second step  will be a 0% hit of available because everything
has already been accelerated in the first step. In this second step
everything will be build from scratch and the local sstate-cache
populadted with the artifacts.

This will improve and significantly reduce the load generated on
the sstate-mirror http server by our CI builds.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>
(cherry picked from commit 68bf73225c0e6e37a3899e0c50a46cf207c87234)
Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>